kodaze/requirements.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.3
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-dbbackup==4.0.1
django-debug-toolbar==3.6.0
django-extensions==3.1.5
django-filter==21.1
django-ninja==0.19.1
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
flower==1.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
humanize==4.2.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.10.2
PyJWT==1.7.1
pyparsing==3.0.9
PyPDF2==1.28.3
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.1.4
reportlab==3.6.9
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0